services/worker/requirements.txt
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
pgvector==0.3.6
celery[redis]==5.4.0
redis==5.2.1
boto3==1.36.4
minio==7.2.12
python-pptx==1.0.2
pypdf==5.1.0
python-docx==1.1.2
openpyxl==3.1.5
numpy==1.26.4
pydantic==2.10.4
pydantic-settings==2.7.1
